Sets dynamic base URL
Updates the base URL configuration to dynamically adapt to the server environment.

This allows the application to automatically adjust its base URL based on the protocol (HTTP or HTTPS) and hostname, eliminating the need for manual configuration changes when deploying to different environments.

Removes the old home controller and adds a new one with only the essential functions.
Updates view to display comments correctly.